Quick and dirty demo, with the actual challenge to be plugged in
access.log is a symlink to /tmp/access.log
index.php appends a line to access.log on every load

The log is displayed, in reverse chronological order (last written line
is at the top). The log is refreshed every one second by loading it from
the server via XHTMLHttpRequest (we could do this via WebSockets, if
the log file were large, bandwidth/latency were premium).

Clicking any of the three shapes of the top of the page makex a
XMLHttpRequest to index.php, which is then logged in access.log.